password_security_system/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
cryptography==42.0.7
argon2-cffi==23.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
pyotp==2.9.0
zxcvbn==4.4.28
qrcode[pil]==7.4.2
rapidfuzz==3.9.3
python-multipart==0.0.9
